+++ src/command/group/business/useCases/groupByFileDateCreatedUseCase.ts
@@ -8,25 +8,29 @@ export class GroupByFileDateCreatedUseCase {
 
     constructor() { }
 
-    group(files: Array<FileWrapper>): void {
+    group(files: Array<FileWrapper>, arg: string): void {
         files.forEach(file => {
-            this._moveFileToItsFolder(file)
+            this._moveFileToItsFolder(file, arg)
         })
     }
 
-    private _moveFileToItsFolder(file: FileWrapper): void {
+    private _moveFileToItsFolder(file: FileWrapper, folderFormat: string): void {
 
-        const folder = this._getFolderFromFileDate(file)
+        const folder = this._getFolderFromFileDate(file, folderFormat)
 
         file.pathNew = path.join(file.pathNew, folder)
     }
 
-    private _getFolderFromFileDate(file: FileWrapper): string {
+    private _getFolderFromFileDate(file: FileWrapper, folderFormat: string): string {
 
-        if(!file.stats) return 'no-date'
+        if (!file.stats) return 'no-date'
 
         const modifiedTime: Date = file.stats.mtime
-        return `${modifiedTime.getFullYear()}-${modifiedTime.getMonth() + 1}-${modifiedTime.getDate()}`
+
+        return folderFormat
+            .replace('YYYY', `${modifiedTime.getFullYear()}`)
+            .replace('MM', `${modifiedTime.getMonth() + 1}`)
+            .replace('DD', `${modifiedTime.getDate()}`)
     }
 
 }