+++ src/command/group/presentation/groupCommandDescription.ts
@@ -1,10 +1,12 @@
 import { Command } from '@commander-js/extra-typings'
 import { Service } from 'typedi'
 import { GroupFiles } from '../business/groupFiles'
-import { GroupOptions } from '../business/groupOptions'
+import { ErrorControlled } from '../../../common/errors';
 
 @Service()
 export class GroupCommandDescription {
+
+    dateFormatRegex = /(YYYY|MM|DD)/;
 
     constructor(
         private readonly _groupFiles: GroupFiles,
@@ -19,11 +21,28 @@ export class GroupCommandDescription {
         program
             .command('group')
             .summary('Group files depending on flags')
-            .option('-d, --date-created', 'By date created')
+            .option('-d, --date-created <format>', 'By date created', undefined)
             .option('-e, --extension', 'By extension type')
-            .action((options: GroupOptions) => {
+            .action((options) => {
+                this._validateInput(options)
                 return this._groupFiles.execute({ path: executionPath, specificOptions: options })
             })
 
     }
+
+    private _validateInput(options: {
+        dateCreated: string | boolean | [] | string[],
+        extension?: true | undefined
+    }) {
+        this._validateDateCreated(options.dateCreated)
+    }
+
+    private _validateDateCreated(dateCreated: string | boolean | [] | string[]) {
+        if (dateCreated == undefined) return
+        if (typeof dateCreated != 'string') throw new ErrorControlled('Type of dateCreated param is not valid')
+        if (!dateCreated.match(this.dateFormatRegex)) throw new ErrorControlled('Date format parameter string must contain YYYY, MM or DD')
+    }
+
 }
+
+