+++ CHANGELOG.md
@@ -1,13 +1,31 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## [0.2.0] - Pending
 
+### Added
+- Command to find duplicated files **findDuplicates**
+
+### Changed
+- Added format parameter to group by date option
+- Controlled errors are now shown as [WARN] on logs
+
+### Fixed
+- Error when reading a directory with lots of files
 
 ## [0.1.1] - 25-01-2023
+
+### Bumped version
+
+### Changed
+- Base command is now **fizard**
+
+## [0.1.0] - 25-01-2023
 ### Added
 - New commands:
-    1. group - groups files based on date modified or/and file type
-    2. structure - creates a structure of folders with a given structure
-    3. destructure - removes all folder structure and moves all files where the command is executed
+    1. **group** - groups files based on date modified or/and file type
+    2. **structure** - creates a structure of folders with a given structure
+    3. **destructure** - removes all folder structure and moves all files where the command is executed
 
-[0.1.0]: https://github.com/JoseLu-Dev/file-cli/compare/0.0.0...0.1.1
+[0.1.0]: https://github.com/JoseLu-Dev/fizard/compare/v0.0.0...v0.1.0
+[0.1.1]: https://github.com/JoseLu-Dev/fizard/compare/v0.1.0...v0.1.1
+[0.2.0]: https://github.com/JoseLu-Dev/fizard/compare/v0.1.1...v0.2.0